Add authorization for reservations
