+++ Gemfile
@@ -51,3 +51,4 @@ gem 'rack-cors'
 gem 'devise'
 gem 'devise-jwt', '~> 0.11.0'
 gem 'jsonapi-serializer'
+gem 'cancancan'
