+++ app/controllers/api/v1/reservations_controller.rb
@@ -1,8 +1,9 @@
 module Api
   module V1
     class ReservationsController < ApplicationController
+      load_and_authorize_resource
       before_action :set_reservation, only: %i[show update destroy]
-
+      
       def index
         @reservations = current_user.reservations
         render json: @reservations
