+++ app/controllers/application_controller.rb
@@ -1,6 +1,6 @@
 class ApplicationController < ActionController::API
     before_action :configure_permitted_parameters, if: :devise_controller?
-    before_action :authenticate_user!
+    before_action :authenticate_api_v1_user!
     protected
     def configure_permitted_parameters
       devise_parameter_sanitizer.permit(:sign_up, keys: %i[name email password password_confirmation])
